Impliment rune-proof space finder.
These two space finder functions look for spaces and if possible return
the index of the space.  If not they return the maximum extent to which
they searched with status flag == false.

In addition, if they are mid codepoint. they step inwards to ensure that
only complete codepoints are sniped.